requirements.txt
fastapi
uvicorn
pandas
networkx
plotly
streamlit
pyvis
python-louvain
google-generativeai
python-dotenv